split read verb handler logic into abstract and concrete classes
